minor renaming and requirements.txt check
