+++ requirements.txt
@@ -1,5 +1,6 @@
-pytorch-lightning == 1.9.5
-torch == 1.13.1
-torchmetrics == 0.11.4
-torchvision == 0.14.1
+pytorch-lightning==1.9.5
+torch==1.13.1
+torchmetrics==0.11.4
+torchvision==0.14.1
 pandas==2.0.2
+numpy==1.26.4
