--- src/cdh1_cancer_dis/structs.py
@@ -1,10 +1,0 @@
-from typing import Dict, NamedTuple
-from torch import Tensor
-
-
-class ModelOutput(NamedTuple):
-    """Torch scriptable model output."""
-
-    backbone_embedding: Tensor
-    heads_activations: Dict[str, Tensor]
-    attn_scores: Tensor
