+++ src/cdh1_cancer_res/structs.py
@@ -1,0 +1,10 @@
+from typing import Dict, NamedTuple
+from torch import Tensor
+
+
+class ModelOutput(NamedTuple):
+    """Torch scriptable model output."""
+
+    backbone_embedding: Tensor
+    heads_activations: Dict[str, Tensor]
+    attn_scores: Tensor
